Added Router to the Navigation Bar
In order to go to the other pages on the navbar or any other page in the app, I used react router to achieve this and therefore I added tht functionality to the nav bar. Also the problem with the app taking long to load is caused by the POB_Cover.svg being too big(22MB), I used JPEG file instead.